a territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	char d;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(d));
-	printf("\nSize of an int: %lu byte(s)\n", sizeof(a));
-	printf("\nSize of a long int: %lu byte(s)\n", sizeof(b));
-	printf("\nSize of a long long int: %lu byte(s)\n", sizeof(c));
-	printf("\nSize of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("\nSize of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("\nSize of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("\nSize of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("\nSize of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
